Add to intro, model, and scheme sections
In the introduction, I added some description to the platelet priming
section and rolling model lit review. In the model chapter, I transferred
the description of the stochastic algorithm from the Appendix to the
model section and updated some of the description of my binning
procedure. More updates will probably be necessary to fully update
this description.