Update files in the oral document folder
In the intro, I've expanded a bit on the drawbacks of AD models.
In the model, I added some background explanation of the modeling
motivation, and some of the assumptions involved in this model.
I expanded a lot on the motivation of the stochastic model, and
its advantage over the deterministic model. Finally, I converted
the existing tables in the document to use the booktabs package.