Fit model parameters to sample data from stochastic simulation
This is a large update that fits both the adiabatic reduction model
and the full PDE model to simulated data. The PDFs generated by the
models with fitted parameters are plotted against a histogram of the
data. The CDFs are also plotted against the ECDF from the data.

A bootstrapping procedure was added as well, but is not yet fully
tested.